Fix stack buffer overflow in tracker handling
commit 06f10c88f557616dac59b4721d333b190ba7b0f3 upstream.

`memcpy(a, b, strlen(b));` is just a complicated way of writing
`strcpy(a, b);`, but even more dangerous since it omits the null
terminator.  `strcpy` on untrusted data is always unsafe, and this data
does not appear to be checked before use.  There is no need to copy the
data before using it, so switch to using it in place.  This eliminates
the security problem and makes the code slightly smaller.

Fixes: 730879d733700ab2c42bb8b9d7c6217ac84c5d96 ("Updated handling of data sent by tracker. ...")